move populating expected responses to when test suite is loaded into library
