+++ internal/app/connectconformance/test_case_library.go
@@ -24,6 +24,7 @@ import (
 	conformancev2 "connectrpc.com/conformance/internal/gen/proto/go/connectrpc/conformance/v2"
 	"github.com/bufbuild/protoyaml-go"
 	"google.golang.org/protobuf/proto"
+	"google.golang.org/protobuf/types/known/anypb"
 )
 
 //nolint:gochecknoglobals
@@ -225,9 +226,191 @@ func parseTestSuites(testFileData map[string][]byte) (map[string]*conformancev2.
 		if err := opts.Unmarshal(data, suite); err != nil {
 			return nil, ensureFileName(err, testFilePath)
 		}
+		for _, testCase := range suite.TestCases {
+			err := populateExpectedResponse(testCase)
+			if err != nil {
+				return nil, fmt.Errorf("%s: failed to compute expected response for test case %q: %w",
+					testFilePath, testCase.Request.TestName, err)
+			}
+		}
 		allSuites[testFilePath] = suite
 	}
 	return allSuites, nil
+}
+
+// populateExpectedResponse populates the response we expected to get back from the server
+// by examining the requests we sent.
+func populateExpectedResponse(testCase *conformancev2.TestCase) error {
+	// If an expected response was already provided, return and use that.
+	// This allows for overriding this function with explicit values in the yaml file.
+	if testCase.ExpectedResponse != nil {
+		return nil
+	}
+	// TODO - This is just a temporary constraint to protect against panics for now.
+	// Eventually, we want to be able to test client and bidi streams where there are no request messages.
+	// The potential plan is for server impls to produce (and the code below to expect) a single response
+	// message in this situation, where the response data value is some fixed string (such as "no response definition")
+	// and whose request info will still be present, but we expect it to indicate zero request messages.
+	if len(testCase.Request.RequestMessages) == 0 {
+		return errors.New("at least one request is required")
+	}
+
+	switch testCase.Request.StreamType {
+	case conformancev2.StreamType_STREAM_TYPE_FULL_DUPLEX_BIDI_STREAM,
+		conformancev2.StreamType_STREAM_TYPE_HALF_DUPLEX_BIDI_STREAM,
+		conformancev2.StreamType_STREAM_TYPE_SERVER_STREAM:
+		return populateExpectedStreamResponse(testCase)
+
+	case conformancev2.StreamType_STREAM_TYPE_UNARY,
+		conformancev2.StreamType_STREAM_TYPE_CLIENT_STREAM:
+		return populateExpectedUnaryResponse(testCase)
+
+	case conformancev2.StreamType_STREAM_TYPE_UNSPECIFIED:
+		return errors.New("stream type is required")
+	default:
+		return fmt.Errorf("stream type %s is not supported", testCase.Request.StreamType)
+	}
+}
+
+// populates the expected response for a unary test case.
+func populateExpectedUnaryResponse(testCase *conformancev2.TestCase) error {
+	req := testCase.Request.RequestMessages[0]
+	// First, find the response definition that the client instructed the server to return
+	concreteReq, err := req.UnmarshalNew()
+	if err != nil {
+		return err
+	}
+	type unaryResponseDefiner interface {
+		GetResponseDefinition() *conformancev2.UnaryResponseDefinition
+	}
+
+	definer, ok := concreteReq.(unaryResponseDefiner)
+	if !ok {
+		return fmt.Errorf("%T is not a unary test case", concreteReq)
+	}
+
+	// TODO - Need to define this better in the protos and tests as to how services should
+	// behave if no responses are specified. The behavior right now differs for unary vs. streaming
+	// If no responses are specified for unary, the service will still return a response with the
+	// request information inside (but none of the response information since it wasn't provided)
+	// But streaming endpoints don't return a single response and instead return responses via sending
+	// on a stream. But if no responses are specified in the request, the streams don't send anything outbound
+	// so there's no way to relay this to a client. So right now, streaming endpoints simply expect an empty
+	// ClientResponseResult if no response definition is provided
+	def := definer.GetResponseDefinition()
+	if def == nil {
+		testCase.ExpectedResponse = &conformancev2.ClientResponseResult{
+			Payloads: []*conformancev2.ConformancePayload{
+				{
+					RequestInfo: &conformancev2.ConformancePayload_RequestInfo{
+						RequestHeaders: testCase.Request.RequestHeaders,
+						Requests:       testCase.Request.RequestMessages,
+					},
+				},
+			},
+		}
+		return nil
+	}
+
+	// Server should have echoed back all specified headers and trailers
+	expected := &conformancev2.ClientResponseResult{
+		ResponseHeaders:  def.ResponseHeaders,
+		ResponseTrailers: def.ResponseTrailers,
+	}
+
+	switch respType := def.Response.(type) {
+	case *conformancev2.UnaryResponseDefinition_Error:
+		// If an error was specified, it should be returned in the response
+		expected.Error = respType.Error
+	case *conformancev2.UnaryResponseDefinition_ResponseData, nil:
+		// If response data was specified for the response (or nothing at all),
+		// the server should echo back the request message and headers in the response
+		payload := &conformancev2.ConformancePayload{
+			RequestInfo: &conformancev2.ConformancePayload_RequestInfo{
+				RequestHeaders: testCase.Request.RequestHeaders,
+				Requests:       testCase.Request.RequestMessages,
+			},
+		}
+		// If response data was specified for the response, it should be returned
+		if respType, ok := respType.(*conformancev2.UnaryResponseDefinition_ResponseData); ok {
+			payload.Data = respType.ResponseData
+		}
+		expected.Payloads = []*conformancev2.ConformancePayload{payload}
+	default:
+		return fmt.Errorf("provided UnaryRequest.Response has an unexpected type %T", respType)
+	}
+
+	testCase.ExpectedResponse = expected
+	return nil
+}
+
+// populates the expected response for a streaming test case.
+func populateExpectedStreamResponse(testCase *conformancev2.TestCase) error {
+	req := testCase.Request.RequestMessages[0]
+	// First, find the response definition that the client instructed the
+	// server to return
+	concreteReq, err := req.UnmarshalNew()
+	if err != nil {
+		return err
+	}
+	type streamResponseDefiner interface {
+		GetResponseDefinition() *conformancev2.StreamResponseDefinition
+	}
+
+	definer, ok := concreteReq.(streamResponseDefiner)
+	if !ok {
+		return fmt.Errorf(
+			"TestCase %s contains a request message of type %T, which is not a streaming request",
+			testCase.Request.TestName,
+			concreteReq,
+		)
+	}
+
+	def := definer.GetResponseDefinition()
+	if def == nil {
+		testCase.ExpectedResponse = &conformancev2.ClientResponseResult{}
+		return nil
+	}
+
+	// Server should have echoed back all specified headers, trailers, and errors
+	expected := &conformancev2.ClientResponseResult{
+		ResponseHeaders:  def.ResponseHeaders,
+		ResponseTrailers: def.ResponseTrailers,
+		Error:            def.Error,
+	}
+
+	// There should be one payload for every ResponseData the client specified
+	expected.Payloads = make([]*conformancev2.ConformancePayload, len(def.ResponseData))
+
+	for idx, data := range def.ResponseData {
+		expected.Payloads[idx] = &conformancev2.ConformancePayload{
+			Data: data,
+		}
+		switch testCase.Request.StreamType { //nolint:exhaustive
+		case conformancev2.StreamType_STREAM_TYPE_SERVER_STREAM,
+			conformancev2.StreamType_STREAM_TYPE_HALF_DUPLEX_BIDI_STREAM:
+			// For server streams and half duplex bidi streams, all request information
+			// specified should only be echoed back in the first response
+			if idx == 0 {
+				expected.Payloads[idx].RequestInfo = &conformancev2.ConformancePayload_RequestInfo{
+					RequestHeaders: testCase.Request.RequestHeaders,
+					Requests:       testCase.Request.RequestMessages,
+				}
+			}
+		case conformancev2.StreamType_STREAM_TYPE_FULL_DUPLEX_BIDI_STREAM:
+			// For a full duplex stream, the first request should be echoed back in the first
+			// payload. The second should be echoed back in the second payload, etc. (i.e. a ping pong interaction)
+			expected.Payloads[idx].RequestInfo = &conformancev2.ConformancePayload_RequestInfo{
+				// RequestHeaders: testCase.Request.RequestHeaders,
+				Requests: []*anypb.Any{testCase.Request.RequestMessages[idx]},
+			}
+			if idx == 0 {
+				expected.Payloads[idx].RequestInfo.RequestHeaders = testCase.Request.RequestHeaders
+			}
+		}
+	}
+	testCase.ExpectedResponse = expected
+	return nil
 }
 
 func generateTestCasePrefix(suite *conformancev2.TestSuite, cfgCase configCase) []string {
